requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
openai==2.8.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
python-dotenv
django-storages
google-cloud-storage
gunicorn
